Write metadata in front + pass it along (useful if the song is changed)
