+++ plugins/downloader/back.js
@@ -1,8 +1,6 @@
-const { writeFileSync } = require("fs");
 const { join } = require("path");
 
-const ID3Writer = require("browser-id3-writer");
-const { dialog, ipcMain } = require("electron");
+const { dialog } = require("electron");
 
 const getSongInfo = require("../../providers/song-info");
 const { injectCSS, listenAction } = require("../utils");
@@ -39,35 +37,6 @@ function handle(win) {
 			default:
 				console.log("Unknown action: " + action);
 		}
-	});
-
-	ipcMain.on("add-metadata", (event, filePath, songBuffer) => {
-		let fileBuffer = songBuffer;
-
-		try {
-			const writer = new ID3Writer(songBuffer);
-			if (metadata.image) {
-				const coverBuffer = metadata.image.toPNG();
-
-				// Create the metadata tags
-				writer
-					.setFrame("TIT2", metadata.title)
-					.setFrame("TPE1", [metadata.artist])
-					.setFrame("APIC", {
-						type: 3,
-						data: coverBuffer,
-						description: "",
-					});
-				writer.addTag();
-			}
-			fileBuffer = Buffer.from(writer.arrayBuffer);
-		} catch (error) {
-			sendError(win, error);
-		}
-
-		writeFileSync(filePath, fileBuffer);
-		// Notify the youtube-dl file
-		event.reply("add-metadata-done");
 	});
 }
